port changes from ezmb1 up to release 6.2.1

+++ Command/MigrationCommand.php
@@ -11,7 +11,8 @@ use Kaliop\eZMigrationBundle\API\Value\MigrationDefinition;
 use Symfony\Component\Console\Question\ConfirmationQuestion;
 
 /**
- * Command to execute the available migration definitions.
+ * Command to manipulate the available migrations / migration definitions.
+ * @todo should we split off the actions used to manipulate migration defs into a separate command `MigrationDefinition` ?
  */
 class MigrationCommand extends AbstractCommand
 {
@@ -32,16 +33,29 @@ class MigrationCommand extends AbstractCommand
             ->addOption('info', null, InputOption::VALUE_NONE, "Get info about the specified migration.")
             ->addOption('add', null, InputOption::VALUE_NONE, "Add the specified migration definition.")
             ->addOption('skip', null, InputOption::VALUE_NONE, "Mark the specified migration as skipped.")
+            ->addOption('fail', null, InputOption::VALUE_NONE, "Mark the specified migration as failed.")
             ->addOption('no-interaction', 'n', InputOption::VALUE_NONE, "Do not ask any interactive question.")
             ->addArgument('migration', InputArgument::REQUIRED, 'The migration to add/skip (filename with full path) or detail/delete (plain migration name).', null)
             ->setHelp(<<<EOT
-The <info>kaliop:migration:migration</info> command allows you to manually delete migrations versions from the migration table:
+The <info>kaliop:migration:migration</info> command allows you to manually manage migrations.
 
-    <info>./ezpublish/console kaliop:migration:migration --delete migration_name</info>
+To see detailed information about a migration or migration definition:
 
-As well as manually adding migrations to the migration table:
+    <info>php bin/console kaliop:migration:migration --info migration_name</info>
 
-    <info>./ezpublish/console kaliop:migration:migration --add /path/to/migration_definition</info>
+    <info>php bin/console kaliop:migration:migration --info /path/to/migration_definition.yml</info>
+
+To remove a migration from the migration table, or mark it as failed:
+
+    <info>php bin/console kaliop:migration:migration --delete migration_name</info>
+
+    <info>php bin/console kaliop:migration:migration --fail migration_name</info>
+
+To manually add migration definitions to the migration table, or marking them as skipped:
+
+    <info>php bin/console kaliop:migration:migration --add /path/to/migration_definition</info>
+
+    <info>php bin/console kaliop:migration:migration --skip /path/to/migration_definition</info>
 EOT
             );
     }
@@ -58,8 +72,9 @@ EOT
         $this->setOutput($output);
         $this->setVerbosity($output->getVerbosity());
 
-        if (!$input->getOption('add') && !$input->getOption('delete') && !$input->getOption('skip') && !$input->getOption('info')) {
-            throw new \InvalidArgumentException('You must specify whether you want to --add, --delete, --skip or --info the specified migration.');
+        if (!$input->getOption('add') && !$input->getOption('delete') && !$input->getOption('skip') && 
+            !$input->getOption('info') && !$input->getOption('fail')) {
+            throw new \InvalidArgumentException('You must specify whether you want to --add, --delete, --skip, --fail or --info the specified migration.');
         }
 
         $migrationService = $this->getMigrationService();
@@ -68,6 +83,7 @@ EOT
         if ($input->getOption('info')) {
             $output->writeln('');
 
+            /// @todo if we are passed a path, we could give the user a more specific warning than what we get back from the storage layer
             $migration = $migrationService->getMigration($migrationNameOrPath);
             if ($migration == null) {
                 throw new \InvalidArgumentException(sprintf('The migration "%s" does not exist in the migrations table.', $migrationNameOrPath));
@@ -111,8 +127,10 @@ EOT
             $output->writeln('Definition md5: <info>' . $migration->md5 . '</info>');
 
             if ($migration->path != '') {
-                // q: what if we have a loader which does not work with is_file? We could probably remove this check...
-                if (is_file($migration->path)) {
+                // q: what if we have a loader which does not work with is_file? We could probably remove this check
+                // or, better, add a method `$migrationService->migrationDefinitionExists` / use a specific exception in getMigrationsDefinitions...
+                // Note also that $migration->path can not be used as is, as it is usually relative to the app's root dir
+                //if (is_file($migration->path)) {
                     try {
                         $migrationDefinitionCollection = $migrationService->getMigrationsDefinitions(array($migration->path));
                         if (count($migrationDefinitionCollection)) {
@@ -133,9 +151,9 @@ EOT
                         /// @todo one day we should be able to limit the kind of exceptions we have to catch here...
                         $output->writeln('Definition parsing error: <error>' . $e->getMessage() . '</error>');
                     }
-                } else {
-                    $output->writeln('Definition error: <error>The migration definition file can not be found any more</error>');
-                }
+                //} else {
+                //    $output->writeln('Definition error: <error>The migration definition file can not be found any more</error>');
+                //}
             }
 
             $output->writeln('');
@@ -174,19 +192,28 @@ EOT
                 }
 
                 $migrationService->addMigration($migrationDefinition);
-                $output->writeln('<info>Added migration' . $migrationDefinition->path . '</info>');
+                $output->writeln('<info>Added migration ' . $migrationDefinition->path . '</info>');
             }
 
             return 0;
         }
 
-        if ($input->getOption('delete')) {
+        if ($input->getOption('delete') || $input->getOption('fail')) {
+            /// @todo if we are passed a path, we could give the user a more specific warning than what we get back from the storage layer
             $migration = $migrationService->getMigration($migrationNameOrPath);
             if ($migration == null) {
                 throw new \InvalidArgumentException(sprintf('The migration "%s" does not exist in the migrations table.', $migrationNameOrPath));
             }
 
-            $migrationService->deleteMigration($migration);
+            if ($input->getOption('delete')) {
+                $migrationService->deleteMigration($migration);
+            } else {
+                $errorMessage = 'Manually failed on ' . date("Y-m-d H:i:s");
+                if ($migration->executionError != '') {
+                    $errorMessage .= ". Previous notes: " . $migration->executionError;
+                }
+                $migrationService->failMigration($migration, $errorMessage);
+            }
 
             return 0;
         }
@@ -202,7 +229,7 @@ EOT
 
             foreach ($migrationDefinitionCollection as $migrationDefinition) {
                 $migrationService->skipMigration($migrationDefinition);
-                $output->writeln('<info>Migration' . $migrationDefinition->path . ' marked as skipped</info>');
+                $output->writeln('<info>Migration ' . $migrationDefinition->path . ' marked as skipped</info>');
             }
 
             return 0;
